Make `thonny_seal` to a separate module
This is necessary so that we can make `thonnycontrib` a namespace
module. Otherwise, if we put `thonny_seal` beneath `thonnycontrib` mypy
will complain that it can access the module from two different paths.
