+++ thonny_seal/py.typed
@@ -1,0 +1,1 @@
+# Marker file for PEP 561. The mypy package uses inline types.
