--- thonnycontrib/__init__.py
@@ -1,1 +1,0 @@
-"""Provide plug-ins for Thonny."""
